Refactorizacion de las funciones en mi base de datos 9
