+++ src/services/sale/getSalesLocal/getSalesLocal.ts
@@ -2,7 +2,7 @@ import { Sale } from '../../../../type/Sale';
 import { getDB, Tx } from '../../../../app/database/database';
 import * as SQLite from 'expo-sqlite';
 
-export const getSaleLocal = async (txn?: Tx): Promise<Sale[]> => {
+export const getSalesLocal = async (txn?: Tx): Promise<Sale[]> => {
   const db = getDB();
   const database: Tx | SQLite.SQLiteDatabase = txn ?? db;
   const saleList = await database.getAllAsync<Sale>(`SELECT id, name, date, status FROM sale;`);
